+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New features:
 
 Behaviour changes:
  * adjust behaviour of ProcessPreBlock callback (#129)
+ * (*DBFT).Header() and (*DBFT).PreHeader() are moved to (*Context) receiver (#133)
 
 Improvements:
  * minimum required Go version is 1.22 (#122, #126)
@@ -16,6 +17,7 @@ Bugs fixed:
  * context-bound PreBlock and PreHeader are not reset properly (#127)   
  * PreHeader is constructed instead of PreBlock to create PreCommit message (#128)
  * enable anti-MEV extension with respect to the current block index (#132)
+ * (*Context).PreBlock() method returns PreHeader instead of PreBlock (#133)
 
 ## [0.3.0] (01 August 2024)